Implement content "update" command (#339)
* Modify content-manager generated command following the issue indications

* Apply changes to validate update command

Includes update option to parse correctly update command by the name update at openapi.yml and Action.java (new FetchConfigCommand type class needed). Rest modified to test. Still unfinished

* Update the Content Manager command model

Add validation of this command type in the Command Manager
Add the new command to the technical documentation and OpenAPI spec.

* Correct minor errors

* Update command structure

* Create testing for Update Content

* Correct consistency and typos

* Move POST command handling to the client class

* Undo updatecontentcommand modification

The validation was thought to be ordered but it is not

* Include Command Manager plugin to the Content Manager's build.gradle

* Fix plugins dependencies

* Refactor

* Remove unused code

* Add documentation

* Comment out zipArchive for the Command Manager

* Fix typo in docs

---------

Signed-off-by: Álex Ruiz <alejandro.ruiz.becerra@wazuh.com>
Co-authored-by: Malena Casas <malena.casas@wazuh.com>
Co-authored-by: Álex Ruiz <alejandro.ruiz.becerra@wazuh.com>